🔧 chore: Update test configuration and gitignore
- Add local data directories to .gitignore
- Update .env.test with CWD-based local store configuration
- Move Bilibili video re-encoding test to appropriate test file
