+++ .env.test
@@ -3,7 +3,8 @@ LOG_LEVEL=INFO
 FASTAPI_RELOAD=false
 
 NICKNAME='["BOT"]'
-
+# 使用当前工作目录作为数据存储目录，以下数据目录配置项默认值将会对应变更
+LOCALSTORE_USE_CWD=true
 
 #r_bili_ck="SESSDATA=jdkcnsjkxbdj"
 r_ytb_ck="a=a;b=b;c=c"
