+++ .gitignore
@@ -151,3 +151,7 @@ uv.lock
 !.vscode/launch.json
 !.vscode/extensions.json
 !.vscode/*.code-snippets
+
+cache
+data
+config
